Revert "Merge remote-tracking branch 'origin/master'"
This reverts commit d83537809fdd5997378958c9103d1072b618c24e, reversing
changes made to 3626d29a2eb4f93c26c715a1d330061608e335ef.
