+++ TeamCode/build.gradle
@@ -33,9 +33,8 @@ dependencies {
     implementation project(':FtcRobotController')
 
     implementation 'org.ftclib.ftclib:core:2.1.1'
-//    implementation "com.acmerobotics.roadrunner:ftc:0.1.14"
+    implementation "com.acmerobotics.roadrunner:ftc:0.1.14"
     implementation "com.acmerobotics.roadrunner:core:1.0.0"
     implementation "com.acmerobotics.roadrunner:actions:1.0.0"
     implementation "com.acmerobotics.dashboard:dashboard:0.4.16"
-    implementation "com.github.jdhs-ftc:road-runner-ftc-otos:8e83c17ff7"
 }
